# docs/openapi.json
{
  "components": {
    "schemas": {
      "AiAnalysisRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "hotspot_id": {
            "title": "Hotspot Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "importance": {
            "title": "Importance",
            "type": "string"
          },
          "is_real": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Real"
          },
          "keyword_mentioned": {
            "title": "Keyword Mentioned",
            "type": "boolean"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "quick_understanding": {
            "items": {
              "type": "string"
            },
            "title": "Quick Understanding",
            "type": "array"
          },
          "raw_response": {
            "additionalProperties": true,
            "title": "Raw Response",
            "type": "object"
          },
          "relevance_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevance Reason"
          },
          "relevance_score": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Relevance Score",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "topic_ideas": {
            "items": {
              "$ref": "#/components/schemas/TopicIdeaRead"
            },
            "title": "Topic Ideas",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "hotspot_id",
          "is_real",
          "relevance_score",
          "relevance_reason",
          "keyword_mentioned",
          "importance",
          "summary",
          "model_name",
          "raw_response",
          "created_at",
          "updated_at"
        ],
        "title": "AiAnalysisRead",
        "type": "object"
      },
...
      "CheckRunCreate": {
        "properties": {
          "trigger_type": {
            "default": "manual",
            "title": "Trigger Type",
            "type": "string"
          }
        },
        "title": "CheckRunCreate",
        "type": "object"
      },
      "CheckRunRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          },
          "failure_count": {
            "title": "Failure Count",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "success_count": {
            "title": "Success Count",
            "type": "integer"
          },
          "trigger_type": {
            "title": "Trigger Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "trigger_type",
          "started_at",
          "finished_at",
          "status",
          "success_count",
          "failure_count",
          "error_summary",
          "created_at",
          "updated_at"
        ],
        "title": "CheckRunRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HotspotClusterResponse": {
        "properties": {
          "cluster_id": {
            "title": "Cluster Id",
            "type": "string"
          },
          "cluster_size": {
            "title": "Cluster Size",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HotspotRead"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "cluster_id",
          "cluster_size",
          "items"
        ],
        "title": "HotspotClusterResponse",
        "type": "object"
      },
      "HotspotRead": {
        "properties": {
          "ai_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AiAnalysisRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "cluster_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Id"
          },
          "cluster_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Version"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "fetched_at": {
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "keyword": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KeywordRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "keyword_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyword Id"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "rank_score": {
            "default": 0,
            "title": "Rank Score",
            "type": "number"
          },
          "raw_payload": {
            "additionalProperties": true,
            "title": "Raw Payload",
            "type": "object"
          },
          "snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snippet"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_id": {
            "title": "Source Id",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "trend_score": {
            "default": 0,
            "title": "Trend Score",
            "type": "number"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "url",
          "source_id",
          "keyword_id",
          "author",
          "snippet",
          "published_at",
          "fetched_at",
          "status",
          "raw_payload",
          "created_at",
          "updated_at"
        ],
        "title": "HotspotRead",
        "type": "object"
      },
...
      "KeywordRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "keyword": {
            "title": "Keyword",
            "type": "string"
          },
          "priority": {
            "default": 0,
            "title": "Priority",
            "type": "integer"
          },
          "query_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Template"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "keyword",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "KeywordRead",
        "type": "object"
      },
...
      "SourceRead": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "name",
          "source_type",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "SourceRead",
        "type": "object"
      },
...
      "TopicIdeaRead": {
        "properties": {
          "angle": {
            "default": "",
            "title": "Angle",
            "type": "string"
          },
          "format": {
            "default": "",
            "title": "Format",
            "type": "string"
          },
          "rationale": {
            "default": "",
            "title": "Rationale",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "TopicIdeaRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Rebuilt FastAPI backend for the self-hosted AI hotspot monitoring MVP.",
    "title": "AI Hotspot Radar API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/check-runs": {
      "get": {
        "operationId": "list_check_runs_api_check_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Check Runs Api Check Runs Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Check Runs",
        "tags": [
          "check-runs"
        ]
      },
      "post": {
        "operationId": "create_check_run_api_check_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckRunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckRunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Check Run",
        "tags": [
          "check-runs"
        ]
      }
    },
    "/api/check-runs/{run_id}": {
      "get": {
        "operationId": "get_check_run_api_check_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckRunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Check Run",
        "tags": [
          "check-runs"
        ]
      }
    },
...
    "/api/hotspots": {
      "get": {
        "operationId": "list_hotspots_api_hotspots_get",
        "parameters": [
          {
            "in": "query",
            "name": "keyword_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword Id"
            }
          },
          {
            "in": "query",
            "name": "source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          },
          {
            "in": "query",
            "name": "importance",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Importance"
            }
          },
          {
            "in": "query",
            "name": "published_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Published From"
            }
          },
          {
            "in": "query",
            "name": "published_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Published To"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "fetched_at_desc",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Hotspots Api Hotspots Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Hotspots",
        "tags": [
          "hotspots"
        ]
      }
    },
    "/api/hotspots/cluster/{cluster_id}": {
      "get": {
        "operationId": "get_hotspot_cluster_api_hotspots_cluster__cluster_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "title": "Cluster Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotspotClusterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Hotspot Cluster",
        "tags": [
          "hotspots"
        ]
      }
    },
    "/api/hotspots/{hotspot_id}": {
      "get": {
        "operationId": "get_hotspot_api_hotspots__hotspot_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "hotspot_id",
            "required": true,
            "schema": {
              "title": "Hotspot Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotspotRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Hotspot",
        "tags": [
          "hotspots"
        ]
      }
    },
    "/api/hotspots/{hotspot_id}/cluster-history": {
      "get": {
        "operationId": "get_hotspot_cluster_history_api_hotspots__hotspot_id__cluster_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "hotspot_id",
            "required": true,
            "schema": {
              "title": "Hotspot Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotspotClusterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Hotspot Cluster History",
        "tags": [
          "hotspots"
        ]
      }
    },
...
  }
}